split player spine logic from shotgun script into player controller script
