+++ Project/Assets/Scripts/PlayerController.cs
@@ -2,30 +2,20 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using Unity.Mathematics;
 using UnityEngine.InputSystem;
 
 [RequireComponent(typeof(PlayerInput))]
 public class PlayerController : MonoBehaviour
-{
-    [HideInInspector]
-    public Rigidbody Rigidbody;
-    public GameObject CameraFollowTarget;
-    public void OnAttack(InputValue value)
-    {
-        // Set the rigidbody velocity opposite the camera directions
-        Rigidbody.velocity -= CameraFollowTarget.transform.forward * 3.0F;
-    }
+{   
+    public Transform cameraFollowTarget;
+    public Transform mainCamera;   
+    public Transform playerModel;
+    public Transform playerSpine;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        if (!TryGetComponent(out Rigidbody))
-        {
-            throw new NullReferenceException("PlayerController did not find Rigidbody component.");
-        }
-        if (CameraFollowTarget == null)
-        {
-            throw new NullReferenceException("PlayerController does not have CameraFollowTarget set.");
-        }
+    void Update(){
+        playerModel.transform.rotation = Quaternion.Euler(0, cameraFollowTarget.transform.rotation.eulerAngles.y, 0);
+        Quaternion lookRotation = Quaternion.LookRotation((mainCamera.position + mainCamera.forward * 10.0f) - playerSpine.position);
+        playerSpine.rotation = math.slerp(playerSpine.rotation, lookRotation, 0.9f);
     }
 }
